feat: prototype notes tools
notes_tool.sh is a backend for the notes. It can list notes and gather
info from them, using a combination of bat (for syntax highliting) and
yq (to read the frontmatter).  It's a bit of a mess right now, and maybe
some functionality (like copying) should be extracted to make the API a
bit cleaner.

notes_entry.sh is an entrypoint suitable for command line use. I've put
this in a separate script because I may add other entrypoints in the
future, i.e. a zle widget. But who knows, maybe it will work it's way
into the tool script eventually.