feat: cache enriched list between calls
Appending tags to each file name during selection is a costly operation,
taking upwards of 5 seconds with ~100 notes.

This is probably a sign that I shouldn't be implementing this with
hacked together scripts - finally an excuse to learn rust?

Until I find time for that, I've implemented a simple caching solution.
The enriched list is saved between calls, and the cache is invalidated
when the content of the notes directory changes.